Adding code of FI with Class and method refs

+++ src/com/tektutorial/funcinterface/Square.java
@@ -3,6 +3,5 @@ package com.tektutorial.funcinterface;
 @FunctionalInterface
 public interface Square {
 
-	int doSquare(int number);	
-	
+	Long doSquare(Long number);	
 }
